requirements.txt
grad_cam==1.5.5
numpy==2.4.3
Pillow==12.1.1
streamlit==1.55.0
torch==2.9.0+cu128
torchvision==0.24.0+cu128
